Docs and secret type
I made the type `TimePicker` secret to have more control over which constructors are allowed, to give us more flexibility in the future.

For example, we don't want the constructor `TimePicker(now(), false)` to work, because we might want to add a third field to our type in the future.

This is something I wish I had done for the other types (Slider, TextField, etc), but it is now too late to change.