Hide ToC contents when hidden
This very small change simply reduce the ToC height to only show the closed book icon when the ToC is hidden. 
This is done to minimize the vertical clutter of the ToC (now that the ToC does not remove its aside status automatically anymore) when the main text is very close to the edge of the browser.
See https://github.com/fonsp/Pluto.jl/issues/1770 for example.